Update test_basic.py
Update tests to use decimal python types

+++ python/pyarrow/tests/parquet/test_basic.py
@@ -19,7 +19,7 @@ from collections import OrderedDict
 import io
 import warnings
 from shutil import copytree
-import decimal
+from decimal import Decimal
 
 import numpy as np
 import pytest
@@ -359,13 +359,13 @@ def test_byte_stream_split():
 
 
 def test_store_decimal_as_integer():
-    arr_decimal_1_9 = pa.array(list(map(float, range(100))), 
+    arr_decimal_1_9 = pa.array(list(map(Decimal, range(100))), 
                                type=pa.decimal128(5,2)
                               )
-    arr_decimal_10_18 = pa.array(list(map(float, range(100))), 
+    arr_decimal_10_18 = pa.array(list(map(Decimal, range(100))), 
                                  type=pa.decimal128(16,9)
                                 )
-    arr_decimal_gt18 = pa.array(list(map(float, range(100))), 
+    arr_decimal_gt18 = pa.array(list(map(Decimal, range(100))), 
                                 type=pa.decimal128(22,2)
                                )
     arr_bool = pa.array([True, False] * 50)
